requirements.txt
argparse
pyfiglet
requests
simple_chalk
python-dotenv